LPS-58307 WikiAttachment view will only accept FileEntry so we can reuse the contributors because both will use FileEntry as return type.
